TABLE/requirements.txt
requests
py-markdown-table